pyproject.toml
[project]
name = "take-home-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.10.0",
    "flask==2.3.3",
    "nltk==3.8.1",
    "numpy>=2.2.3",
    "openai==1.8.0",
    "pandas>=2.2.3",
    "pymupdf==1.22.3",
    "python-docx==0.8.11",
    "python-dotenv==1.0.0",
    "scikit-learn>=1.6.1",
    "tqdm==4.66.1",
    "werkzeug==2.3.7",
]
